fix ciricular deps
