+++ packages/agent-kit/src/agent.ts
@@ -15,8 +15,7 @@ import { serializeError } from "inngest/helpers/errors";
 import { type MinimalEventPayload } from "inngest/types";
 import type { ZodType } from "zod";
 import { createAgenticModelFromAiAdapter, type AgenticModel } from "./model";
-import { NetworkRun } from "./networkRun";
-import { createNetwork } from "./network";
+import { createNetwork, NetworkRun } from "./network";
 import { State, type StateData } from "./state";
 import { type MCP, type Tool } from "./tool";
 import { AgentResult, type Message, type ToolResultMessage } from "./types";
