+++ packages/agent-kit/src/index.ts
@@ -3,7 +3,6 @@ export * from "./agent";
 export * from "./model";
 export * from "./models";
 export * from "./network";
-export * from "./networkRun";
 export * from "./state";
 export * from "./tool";
 export * from "./util";
